Price Points and National Average reports fixed.. Date selection and 0 value cases, price formatting etc

+++ JsPlc.Ssc.PetrolPricing/JsPlc.Ssc.PetrolPricing.Models/ViewModels/NationalAverageReportBrandViewModel.cs
@@ -4,6 +4,6 @@ namespace JsPlc.Ssc.PetrolPricing.Models.ViewModels
     public class NationalAverageReportBrandViewModel
     {
         public string BrandName { get; set; }
-        public double Average { get; set; }
+        public int Average { get; set; }
     }
 }
